backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI/ML相关
diffusers>=0.21.0
transformers>=4.35.0

# 图像处理
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# 数据处理
pydantic>=2.4.0
python-multipart>=0.0.6

# 异步任务处理
celery>=5.3.0
redis>=5.0.0  # 或者使用 RabbitMQ

# 工具库
requests>=2.31.0
aiofiles>=23.2.0
python-dotenv>=1.0.0

# 开发工具
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0

# 日志
loguru>=0.7.0

# 数据库（可选）
sqlalchemy>=2.0.0
alembic>=1.12.0

# 监控（可选）
prometheus-client>=0.18.0
